Create shadow account when person is created
Before a person is created the model will try to create a shadow account in People Aggregator. If this fails, it will rollback the transaction and the person will not be created. You can turn off this behavior at the instance level by setting Person#skip_shadow_account = true. No error checking is done at this time. That's coming next.