Temporarily remove validation for production issue
When registering, everything works fine after a restart. However, for the next registration the avatar validation runs and fails if the user doesn't use an avatar. We need to figure out exactly why this is running on the second time after restarting, but not the first time. The correct behavior is for the validation to not run unless the avatar is present.
(cherry picked from commit 1966179b63d57fbb97c196559dbd73518790606b)